speech-synthesis/local_requirements.txt
# Speech Synthesis Research Pipeline - Local Requirements
# Core ML and Deep Learning
torch>=2.7.1
numpy>=2.2.6
scikit-learn>=1.7.1
scipy>=1.16.1

# Audio Processing
librosa>=0.11.0
soundfile>=0.13.1
pesq>=0.0.4
pystoi>=0.4.1

# Text Processing
phonemizer>=3.3.0
nltk>=3.9.1

# Web Interface
streamlit>=1.47.1

# Visualization
matplotlib>=3.10.5
plotly>=6.2.0
seaborn>=0.13.2

# Data Processing
pandas>=2.3.1
pyyaml>=6.0.2

# Utilities
colorlog>=6.9.0
requests>=2.32.4

# Note: For local installation, you may need to install espeak separately:
# Ubuntu/Debian: sudo apt-get install espeak espeak-data
# macOS: brew install espeak
# Windows: Download from http://espeak.sourceforge.net/download.html
#
# To install all dependencies:
# pip install -r local_requirements.txt
#
# For GPU support (optional):
# Visit https://pytorch.org/get-started/locally/ for CUDA installation instructions